Close bug 23.
The 'reason' for this bug is that '&' is a URL character that indicates
an option. For example:

     /cgi-bin/somescript.cgi?first=option&second=other&etc=something

Now, when the URL contains 'Falu Bil- & Industrifarg' (which, in
English would be 'Falu Car- & Industrycoulor'), the browser thinks
that ' Industrifarg' is an option to 'Falu bil- '. Hence the error...

The fix is to URL encode (exchange speciall URL characters like '=', '&'
etc) with a special character. The '=' character would be '%3D' etc...

The problem with this (URL encode branch and rootdn) is that I use that
(the root and branch DN) in the global config variable to distinguish
variables between branches.

The fix is to create a function - pql_get_define() - that will decode the
DN (if supplied - a branch option) and send the result back to the caller.