Add support for phpQLAdminMXHostAddition:nonPrimaryRcptHosts which is an internal phpQLAdmin way of remembering locally added rcptHosts.
These are domain names that don't exist in the database (as a primary/additional
branch domain name etc) but is still domains the QLC host is listening on (kinda
a backup MX).

Domains not listed as a primary/additional branch domain will not be replicated
when/if doing a manuall replication. This attribute is the rescue (not a NICE
rescue, but still :)